Used iterators + mapping instead of raw for loop

+++ src/branch_deletion_structure.rs
@@ -1,7 +1,7 @@
 #[derive(PartialEq)]
 pub struct BranchDeletionStructure
 {
-    pub(crate) index: i32,
+    pub(crate) index: usize,
     pub(crate) branch_name: String,
     pub(crate) should_be_deleted: bool,
 }
